+++ src/Tiny_Websockets_Generic/internals/ws_common_Ethernet_W5x00.hpp
@@ -9,7 +9,8 @@
   
   Built by Khoi Hoang https://github.com/khoih-prog/Websockets2_Generic
   Licensed under MIT license
-  Version: 1.10.3
+  
+  Version: 1.11.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -20,7 +21,8 @@
   1.10.0  K Hoang      18/12/2021 Supporting case-insensitive headers, according to RFC2616
   1.10.1  K Hoang      26/02/2022 Reduce QNEthernet latency
   1.10.2  K Hoang      14/03/2022 Fix bug when using QNEthernet staticIP. Add staticIP option to NativeEthernet
-  1.10.3  K Hoang      11/04/2022 Use Ethernet_Generic library as default. Support SPI1/SPI2 for RP2040/ESP32
+  1.10.3  K Hoang      11/04/2022 Use Ethernet_Generic library as default. Support SPI1/SPI2 for RP2040
+  1.11.0  K Hoang      08/10/2022 Add support to ESP32 using W5x00 Ethernet
  *****************************************************************************************************************************/
  
 #pragma once
@@ -159,5 +161,21 @@ namespace websockets2_generic
     // OpenSSL Dependent
     #define WSDefaultSecuredTcpClient websockets2_generic::network2_generic::SecuredEthernetTcpClient
     #endif //_WS_CONFIG_NO_SSL      
-              
+
+  #elif ( defined(ESP32) )
+    // Using Ethernet W5x00
+    #warning Using Ethernet for ESP32 in ws_common_Ethernet_W5x00.hpp
+    
+    #define PLATFORM_DOES_NOT_SUPPORT_BLOCKING_READ
+    #define _WS_CONFIG_NO_SSL   true
+    
+    #include <Tiny_Websockets_Generic/network/ESP32_Ethernet_W5x00/ESP32_Ethernet_W5x00_tcp.hpp>
+    #define WSDefaultTcpClient websockets2_generic::network2_generic::EthernetTcpClient
+    #define WSDefaultTcpServer websockets2_generic::network2_generic::EthernetTcpServer
+    
+    #ifndef _WS_CONFIG_NO_SSL
+    // OpenSSL Dependent
+    #define WSDefaultSecuredTcpClient websockets2_generic::network2_generic::SecuredEthernetTcpClient
+    #endif //_WS_CONFIG_NO_SSL
+                  
   #endif